+++ src/pas/plugins/headers/utils.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from pas.plugins.headers.plugins import HeaderPlugin
 from Products.CMFCore.utils import getToolByName
 
 import six
@@ -16,6 +15,10 @@ def get_plugin(context):
     plugin = getattr(pas, PLUGIN_ID, None)
     if plugin is None:
         return
+
+    # Import here to avoid a circular import
+    from pas.plugins.headers.plugins import HeaderPlugin
+
     if not isinstance(plugin, HeaderPlugin):
         return
     return plugin