+++ src/pas/plugins/headers/tests/test_exportimport.py
@@ -119,23 +119,23 @@ class TestImport(ExportImportBaseTestCase):
         self.assertTupleEqual(
             self.plugin.allowed_roles,
             (
-                b'Member',
-                b'Zebra',
+                'Member',
+                'Zebra',
             ),
         )
         self.assertTrue(self.plugin.deny_unauthorized)
         self.assertTupleEqual(
             self.plugin.memberdata_to_header,
             (
-                b'uid|HEADER_uid|lower',
-                b'fullname|HEADER_firstname HEADER_lastname',
+                'uid|HEADER_uid|lower',
+                'fullname|HEADER_firstname HEADER_lastname',
             ),
         )
         self.assertEqual(
-            self.plugin.redirect_url, b'https://maurits.vanrees.org')
-        self.assertTupleEqual(self.plugin.required_headers, (b'uid', b'test'))
-        self.assertEqual(self.plugin.roles_header, b'roles')
-        self.assertEqual(self.plugin.userid_header, b'uid')
+            self.plugin.redirect_url, 'https://maurits.vanrees.org')
+        self.assertTupleEqual(self.plugin.required_headers, ('uid', 'test'))
+        self.assertEqual(self.plugin.roles_header, 'roles')
+        self.assertEqual(self.plugin.userid_header, 'uid')
 
     def test_import_no_file(self):
         """When the import file is not there, nothing should go wrong."""
@@ -169,40 +169,40 @@ class TestImport(ExportImportBaseTestCase):
         import_properties(self._makeContext(json.dumps(settings)))
         self.assertTupleEqual(
             self.plugin.allowed_roles,
-            (b'missile', b'target'),
+            ('missile', 'target'),
         )
         self.assertTrue(self.plugin.deny_unauthorized)
         self.assertTupleEqual(
             self.plugin.memberdata_to_header,
             (
-                b'uid|PROFILE_uid',
-                b'fullname|PROFILE_firstname PROFILE_lastname',
-                b'role|PROFILE_role|lower',
+                'uid|PROFILE_uid',
+                'fullname|PROFILE_firstname PROFILE_lastname',
+                'role|PROFILE_role|lower',
             ),
         )
         self.assertEqual(
-            self.plugin.redirect_url, b'https://maurits.vanrees.org')
+            self.plugin.redirect_url, 'https://maurits.vanrees.org')
         self.assertTupleEqual(
             self.plugin.required_headers,
-            (b'uid', b'role'),
+            ('uid', 'role'),
         )
-        self.assertEqual(self.plugin.roles_header, b'portal_roles')
-        self.assertEqual(self.plugin.userid_header, b'uid')
+        self.assertEqual(self.plugin.roles_header, 'portal_roles')
+        self.assertEqual(self.plugin.userid_header, 'uid')
 
         # Explicitly test the types.
-        # We want string, not unicode: when you save the properties form
-        # in the ZMI, you always get a string.
+        # We want string: when you save the properties form
+        # in the ZMI, you always get a string, in both Py 2 and 3
         # And we want tuples, not lists.
         self.assertIsInstance(self.plugin.allowed_roles, tuple)
-        self.assertIsInstance(self.plugin.allowed_roles[0], six.binary_type)
+        self.assertIsInstance(self.plugin.allowed_roles[0], str)
         self.assertIsInstance(self.plugin.deny_unauthorized, bool)
         self.assertIsInstance(self.plugin.memberdata_to_header, tuple)
-        self.assertIsInstance(self.plugin.memberdata_to_header[0], six.binary_type)
-        self.assertIsInstance(self.plugin.redirect_url, six.binary_type)
+        self.assertIsInstance(self.plugin.memberdata_to_header[0], str)
+        self.assertIsInstance(self.plugin.redirect_url, str)
         self.assertIsInstance(self.plugin.required_headers, tuple)
-        self.assertIsInstance(self.plugin.required_headers[0], six.binary_type)
-        self.assertIsInstance(self.plugin.roles_header, six.binary_type)
-        self.assertIsInstance(self.plugin.userid_header, six.binary_type)
+        self.assertIsInstance(self.plugin.required_headers[0], str)
+        self.assertIsInstance(self.plugin.roles_header, str)
+        self.assertIsInstance(self.plugin.userid_header, str)
 
     def test_import_purge_false(self):
         """Test purge=false."""
@@ -234,7 +234,7 @@ class TestImport(ExportImportBaseTestCase):
         self.assertTupleEqual(self.plugin.memberdata_to_header, ())
         self.assertEqual(self.plugin.redirect_url, '')
         self.assertTupleEqual(self.plugin.required_headers, ())
-        self.assertEqual(self.plugin.userid_header, b'my_uid')
+        self.assertEqual(self.plugin.userid_header, 'my_uid')
 
     def test_import_no_plugin(self):
         """Test that import does not fail when plugin is not there."""
@@ -274,7 +274,7 @@ class TestImport(ExportImportBaseTestCase):
             'userid_header': 'my_uid',
         }
         import_properties(self._makeContext(json.dumps(settings)))
-        self.assertEqual(self.plugin.userid_header, b'my_uid')
+        self.assertEqual(self.plugin.userid_header, 'my_uid')
         marker = object()
         self.assertIs(getattr(self.plugin, 'unknown', marker), marker)
 
@@ -322,7 +322,6 @@ class TestExport(ExportImportBaseTestCase):
         context = self._makeContext()
         export_properties(context)
         # The properties are exported by json.dumps, which returns str in both Python 2 and 3.
-        # self.assertIsInstance(context.get_exported_data(), six.binary_type)
         self.assertIsInstance(context.get_exported_data(), str)
         # json.loads always gives unicode
         self.assertDictEqual(