+++ CHANGES.rst
@@ -2,7 +2,7 @@ Changelog
 =========
 
 
-1.2.1 (unreleased)
+1.3.0 (unreleased)
 ------------------
 
 - We require ``six>=1.12.0`` because we use ``ensure_str``.